// Client/src/app/Model/apprenant.ts
export class Apprenant {
    id: number;
    name: string;
    firstname: string;
    dateNais: Date;
    adresse: string;
    cin: number;
    phone: string;
    role: string;
    email: string;
    password: string;
    roleDesignation: string;
    about: string;
    image: string;

    constructor() {

    }
}